Added more controller & raspberry pi support, added a new mob and a new skill, made many changes and optimizations, and made some quality of life improvements.

+++ core/src/main/java/com/lordzintick/game/skill/SkillType.java
@@ -17,9 +17,9 @@ public class SkillType {
     public final int manaCost;
     public final float cooldown;
     public final Sound castSound;
-    public final BiConsumer<Player, Integer> action;
+    public final BiConsumer<Player, Float> action;
 
-    public SkillType(String displayName, Rarity rarity, List<Text> description, TextureRegion icon, int manaCost, float cooldown, Sound castSound, BiConsumer<Player, Integer> action) {
+    public SkillType(String displayName, Rarity rarity, List<Text> description, TextureRegion icon, int manaCost, float cooldown, Sound castSound, BiConsumer<Player, Float> action) {
         this.displayName = displayName;
         this.rarity = rarity;
         this.description = description;
